update ubi-utils: add new files
git-svn-id: svn://svn.dd-wrt.com/DD-WRT@41254 52c4871e-980c-0410-b1e0-e73912ce01f8

+++ src/router/ubi-utils/misc-utils/serve_image.c
@@ -18,7 +18,10 @@
 #include <crc32.h>
 #include <inttypes.h>
 
+#include <common.h>
+
 #include "mcast_image.h"
+#include "libfec.h"
 
 int tx_rate = 80000;
 int pkt_delay;
@@ -126,8 +129,8 @@ int main(int argc, char **argv)
 	}
 
 	if (st.st_size % erasesize) {
-		fprintf(stderr, "Image size %" PRIu64 " bytes is not a multiple of erasesize %d bytes\n",
-				st.st_size, erasesize);
+		fprintf(stderr, "Image size %lld bytes is not a multiple of erasesize %d bytes\n",
+				(long long)st.st_size, erasesize);
 		exit(1);
 	}
 	image = mmap(NULL, st.st_size, PROT_READ, MAP_PRIVATE, rfd, 0);
